Fix Python 3.10 incompatibilities (#331)
* Fix changes to module locations

* Explicit integer type needed for Qt binding

* Empty commit to re-trigger CI

* Fix CI environments not using declared python version

* Fix docstring issues in Python 3.10

* Update CI to use Python 3.10 instead of 3.9

* Update website building CI job to use Python 3.10

* Fix formatting of "3.10" because YAML is hard

* Fix passing floats to slider position setting

* Add workaround for bug in QGroupBox alignment handling

+++ uwsift/view/rgb_config.py
@@ -171,7 +171,7 @@ class RGBLayerConfigPane(QObject):
         return (slider_val / self._slider_steps) * (valid_max - valid_min) + valid_min
 
     def _create_slider_value(self, valid_min, valid_max, channel_val):
-        return ((channel_val - valid_min) / (valid_max - valid_min)) * self._slider_steps
+        return int((channel_val - valid_min) / (valid_max - valid_min)) * self._slider_steps
 
     def _min_max_for_color(self, rgba: str):
         """
